add: username and status appear in profile area
