+++ react-vite/src/components/Navigation/Navigation.css
@@ -62,6 +62,7 @@ nav a {
 }
 
 #profile-area {
+  position: relative;
   display: flex;
 
   height: 100px;
@@ -97,4 +98,10 @@ nav a {
   color: #a6a6a6;
   font-size: 14px;
   font-weight: 400;
+}
+
+.profile-dropdown {
+  position: absolute;
+  bottom: 0;
+  right: 0;
 }
